changed ResetPlayerAvatar argument from u8 to u16

+++ src/field_player_avatar.c
@@ -2135,7 +2135,7 @@ static bool32 Fishing_StartMinigame(struct Task *task)
     return FALSE;
 }
 
-void ResetPlayerAvatar(u8 gfxId)
+void ResetPlayerAvatar(u16 gfxId)
 {
     struct ObjectEvent *playerObjEvent = &gObjectEvents[gPlayerAvatar.objectEventId];
